Rename "draw" folder to "recording" folder
