+++ src/components/recording/Brush.js
