+++ src/components/recording/BrushThicknessPicker.js
