+++ src/components/recording/BrushThicknessPickerContainer.js
