+++ src/components/recording/ClearButton.js
