+++ src/components/recording/ClearButtonContainer.js
