+++ src/components/recording/ColorLoader.js
