+++ src/components/recording/ColorPicker.js
