+++ src/components/recording/ColorPickerContainer.js
