+++ src/components/recording/DrawViewContainer.js
