+++ src/components/recording/DrawingSidebar.js
