+++ src/components/recording/MouseDrawingHandler.js
